refactor(frontend): relocate theme script
