+++ packages/frontend/src/theme.ts
@@ -8,8 +8,8 @@ import tinycolor from 'tinycolor2';
 import lightTheme from '@@/themes/_light.json5';
 import darkTheme from '@@/themes/_dark.json5';
 import JSON5 from 'json5';
-import { deepClone } from './clone.js';
 import type { BundledTheme } from 'shiki/themes';
+import { deepClone } from '@/utility/clone.js';
 import { globalEvents } from '@/events.js';
 import { miLocalStorage } from '@/local-storage.js';
 import { addTheme, getThemes } from '@/theme-store.js';
